requirements.txt
APScheduler>=3.5.3
certifi>=2018.11.29
chardet>=3.0.4
Click>=7.0
cluster>=1.4.1.post2
coreapi>=2.3.3
coreschema>=0.0.4
coverage>=4.5.2
dj-static>=0.0.6
Django>=2.1.5
django-choices>=1.6.1
django-core>=1.4.1
django-cors-headers>=2.4.0
django-filter>=2.0.0
django-rest-auth>=0.9.3
django-rest-swagger>=2.2.0
django-test>=0.4030
djangorestframework>=3.9.0
djangorestframework-filters>=0.10.2
easydict>=1.9
env>=0.1.0
gunicorn>=19.9.0
idna>=2.7
iohandler>=1.0.6
itypes>=1.1.0
Jinja2>=2.10
lark-parser>=0.5.6
Markdown>=2.6.11
MarkupSafe>=1.1.0
openapi-codec>=1.3.2
order>=0.1.17
pathlib>=1.0.1
Pillow>=5.2.0
popper>=1.1.1
psycopg2==2.7.5
psycopg2-binary>=2.7.6.1
pyaes>=1.6.1
python-crontab>=2.3.5
python-dateutil>=2.7.5
pytz>=2018.9
PyYAML>=4.2b4
renderer>=1.0.3
requests>=2.21.0
scinum>=0.2.7
simplejson>=3.16.0
six>=1.12.0
static3>=0.7.0
structlog>=18.2.0
switch>=1.1.0
typing>=3.6.6
tzlocal>=1.5.1
uritemplate>=3.0.0
urllib3>=1.23
